Updated posts search feature
