+++ client/package.json
@@ -37,5 +37,5 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "https://localhost:5000"
+  "proxy": "http://localhost:5000"
 }
